Major update
we no longer need to add tags for each new exercise we are trying to display.
our js loader will automatically make a unordered list with all the exercises in our json file. Also we need to create the style for <ul> as a class because we dont want the that list to be style like the menu